component file restructure
