+++ src/components/Footer/index.js
@@ -1,0 +1,2 @@
+import Footer from './Footer'
+export default Footer
