+++ src/components/Header/Header.js
@@ -2,8 +2,8 @@ import React from 'react'
 import { Link } from 'gatsby'
 import PropTypes from 'prop-types'
 import posed from 'react-pose'
-import { ShortLine } from './styles/Line'
-import { StyledHeader } from './styles/Header.css'
+import { ShortLine } from '../styles/Line'
+import { StyledHeader } from './Header.css'
 
 const AnimatedHeader = posed(StyledHeader)({
   enter: {
