+++ src/components/Nav/Nav.css.js
@@ -1,5 +1,5 @@
 import styled from 'styled-components'
-import theme from './_theme'
+import theme from '../styles/_theme'
 
 export const StyledNav = styled.div`
   font-family: 'Josefin Sans';
