+++ src/components/Nav/index.js
@@ -1,0 +1,2 @@
+import Nav from './Nav'
+export default Nav
